cleanup page screen
