+++ apps/app/navigation/Navigation.tsx
@@ -7,7 +7,6 @@ import {
   NavigationContainer,
 } from "@react-navigation/native";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import { LocalDevice } from "@serenity-tools/common";
 import {
   Heading,
   Text,
@@ -15,27 +14,21 @@ import {
   useIsDesktopDevice,
   useIsPermanentLeftSidebar,
 } from "@serenity-tools/ui";
-import { useActor, useInterpret } from "@xstate/react";
 import * as Linking from "expo-linking";
 import * as React from "react";
 import { useEffect } from "react";
 import { ColorSchemeName, StyleSheet, useWindowDimensions } from "react-native";
-import { Drawer } from "react-native-drawer-layout";
-import { useSafeAreaInsets } from "react-native-safe-area-context";
 import AccountSettingsSidebar from "../components/accountSettingsSidebar/AccountSettingsSidebar";
-import CommentsSidebar from "../components/commentsSidebar/CommentsSidebar";
 import { HeaderLeft } from "../components/headerLeft/HeaderLeft";
 import NavigationDrawerModal from "../components/navigationDrawerModal/NavigationDrawerModal";
 import { PageHeaderLeft } from "../components/pageHeaderLeft/PageHeaderLeft";
 import Sidebar from "../components/sidebar/Sidebar";
 import WorkspaceSettingsSidebar from "../components/workspaceSettingsSidebar/WorkspaceSettingsSidebar";
-import { PageProvider } from "../context/PageContext";
 import { WorkspaceProvider } from "../context/WorkspaceContext";
 import { useWorkspaceQuery } from "../generated/graphql";
 import { redirectToLoginIfMissingTheActiveDeviceOrSessionKey } from "../higherOrderComponents/redirectToLoginIfMissingTheActiveDeviceOrSessionKey";
 import { useAuthenticatedAppContext } from "../hooks/useAuthenticatedAppContext";
 import { useInterval } from "../hooks/useInterval";
-import { commentsMachine } from "../machines/commentsMachine";
 import {
   RootStackParamList,
   WorkspaceStackParamList,
@@ -84,64 +77,6 @@ const isPhoneDimensions = (width: number) => width < 768;
 
 const drawerWidth = 240;
 
-const PageProviderWrapper: React.FC<{ route: any; navigation: any }> = (
-  props
-) => {
-  const [open, setOpen] = React.useState(false);
-  const { activeDevice } = useAuthenticatedAppContext();
-  const commentsService = useInterpret(commentsMachine, {
-    context: {
-      params: {
-        pageId: props.route.params.pageId,
-        activeDevice: activeDevice as LocalDevice,
-      },
-    },
-  });
-  const [, send] = useActor(commentsService);
-  const isPermanentLeftSidebar = useIsPermanentLeftSidebar();
-  const insets = useSafeAreaInsets();
-
-  return (
-    <PageProvider
-      value={{
-        pageId: props.route.params.pageId,
-        commentsService,
-        setActiveSnapshotAndCommentKeys: (activeSnapshot, commentKeys) => {
-          send({
-            type: "SET_ACTIVE_SNAPSHOT_AND_COMMENT_KEYS",
-            activeSnapshot,
-            commentKeys,
-          });
-        },
-      }}
-    >
-      <Drawer
-        open={open}
-        onOpen={() => setOpen(true)}
-        onClose={() => setOpen(false)}
-        renderDrawerContent={() => {
-          return <CommentsSidebar />;
-        }}
-        drawerType="front"
-        drawerPosition="right"
-        overlayStyle={{
-          display: "none",
-        }}
-        drawerStyle={{
-          width: drawerWidth,
-          marginLeft: isPermanentLeftSidebar ? -drawerWidth : undefined,
-          // necessary to avoid overlapping with the header
-          marginTop: 50 + insets.top,
-          borderLeftWidth: 1,
-          borderLeftColor: tw.color("gray-200"),
-        }}
-      >
-        <PageScreen key={props.route.params.pageId} {...props} />
-      </Drawer>
-    </PageProvider>
-  );
-};
-
 // By remounting the component we make sure that a fresh state machine gets started.
 // As an alternative we could also have an action that resets the state machine,
 // but with all the side-effects remounting seemed to be the stabler choice for now
@@ -151,7 +86,7 @@ const PageScreenWrapper: React.FC<{
   route: any;
   navigation: any;
 }> = (props) => {
-  return <PageProviderWrapper key={props.route.params.pageId} {...props} />;
+  return <PageScreen key={props.route.params.pageId} {...props} />;
 };
 
 function WorkspaceDrawerNavigator(props) {
